+++ CHANGELOG.md
@@ -24,7 +24,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://bitbucket.org/atlassian/report/branches/compare/master%0Drelease-2.3.1
+[Unreleased]: https://bitbucket.org/atlassian/report/branches/compare/master%0Drelease-2.4.0
+
+## [2.4.0] - 2019-01-08
+[2.4.0]: https://bitbucket.org/atlassian/report/branches/compare/release-2.4.0%0Drelease-2.3.1
 
 ### Added
 - Provide builder for EdibleResult.
@@ -33,7 +36,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 [JPERF-338]: https://ecosystem.atlassian.net/browse/JPERF-338
 
 ## [2.3.1] - 2019-01-08
-[2.3.1]: https://bitbucket.org/atlassian/report/branches/compare/release-2.3.1%0Drelease-2.3.1
+[2.3.1]: https://bitbucket.org/atlassian/report/branches/compare/release-2.3.1%0Drelease-2.3.0
 
 ### Fixed
 - Create a JUnit report from exceptions containing '%' character. Fix [JPERF-339].